requirements.txt
# Core dependencies
diffusers>=0.25.0
transformers>=4.36.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
accelerate>=0.27.0
safetensors>=0.4.0
requests>=2.31.0

# PyTorch - will be installed with specific platform command
# On macOS with Apple Silicon, use: 
# pip install torch torchvision

# Testing dependencies
pytest>=7.0.0
httpx>=0.24.0  # Required for TestClient
